+++ src/main/java/de/tum/in/test/api/dynamic/Checkable.java
@@ -1,4 +1,4 @@
-package de.tum.in.test.api.ext;
+package de.tum.in.test.api.dynamic;
 
 public interface Checkable {
 	void check();
