+++ src/main/java/de/tum/in/test/api/dynamic/DynamicClass.java
@@ -1,7 +1,6 @@
-package de.tum.in.test.api.ext;
+package de.tum.in.test.api.dynamic;
 
-import static org.junit.jupiter.api.Assertions.fail;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
