+++ src/main/java/de/tum/in/test/api/dynamic/DynamicConstructor.java
@@ -1,6 +1,6 @@
-package de.tum.in.test.api.ext;
+package de.tum.in.test.api.dynamic;
 
-import static de.tum.in.test.api.ext.DynamicMethod.*;
+import static de.tum.in.test.api.dynamic.DynamicMethod.*;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import java.lang.reflect.Constructor;
