+++ src/main/java/de/tum/in/test/api/dynamic/DynamicField.java
@@ -1,4 +1,4 @@
-package de.tum.in.test.api.ext;
+package de.tum.in.test.api.dynamic;
 
 import static org.junit.jupiter.api.Assertions.fail;
